refactor: Update import paths for interfaces in BattleScene, Pokedex, GameContext, and page
Update the import paths for the Pokemon interface in BattleScene, Pokedex, GameContext, and page files to use the new location in the mainInterface folder.
