+++ frontend/app/interfaces/mainInterface.ts
